main: key_handler.cpp: Remove redundant checks
In single tap event there is no need at all to check whether we are in single bud mode or if the event came from left or right earbud in TWS mode. The event is the same.

For quad tap we do nothing at all in TWS mode for now. So remove the checks for right or left earbud.